+++ payments-ui-core/src/test/java/com/stripe/android/ui/core/elements/AddressControllerTest.kt
@@ -27,7 +27,7 @@ class AddressControllerTest {
                 emailController
             ),
             IbanElement(
-                IdentifierSpec.Generic("iban"),
+                IdentifierSpec.Generic("sepa_debit[iban]"),
                 ibanController
             )
         )