+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/test/core/PlaygroundTestDriver.kt
@@ -1,19 +1,20 @@
 package com.stripe.android.test.core
 
+import android.app.Activity
+import android.app.Application
 import android.graphics.Bitmap
+import android.os.Bundle
 import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.test.core.app.ActivityScenario
+import androidx.test.espresso.Espresso
 import androidx.test.espresso.IdlingPolicies
 import androidx.test.espresso.IdlingRegistry
 import androidx.test.runner.screenshot.Screenshot
 import androidx.test.uiautomator.UiDevice
 import com.google.common.truth.Truth.assertThat
 import com.stripe.android.paymentsheet.PaymentSheetResult
-import com.stripe.android.paymentsheet.example.R
 import com.stripe.android.paymentsheet.example.playground.activity.PaymentSheetPlaygroundActivity
-import com.stripe.android.paymentsheet.viewmodels.TransitionFragmentResource
 import com.stripe.android.test.core.ui.BrowserUI
-import com.stripe.android.test.core.ui.EspressoLabelIdButton
 import com.stripe.android.test.core.ui.EspressoText
 import com.stripe.android.test.core.ui.Selectors
 import org.junit.Assume
@@ -36,9 +37,70 @@ class PlaygroundTestDriver(
     private val basicScreenCaptureProcessor: MyScreenCaptureProcessor,
 ) {
     private var resultValue: String? = null
-    private val callbackLock = Semaphore(1)
     private lateinit var testParameters: TestParameters
     private lateinit var selectors: Selectors
+
+    private val currentActivity = Array<Activity?>(1) { null }
+    private var application: Application? = null
+
+    private val activityLifecycleCallbacks = object : Application.ActivityLifecycleCallbacks {
+        override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {}
+        override fun onActivityStarted(activity: Activity) {}
+        override fun onActivityPaused(activity: Activity) {}
+        override fun onActivityStopped(activity: Activity) {}
+        override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {}
+        override fun onActivityDestroyed(activity: Activity) {}
+        override fun onActivityResumed(activity: Activity) {
+            currentActivity[0] = activity
+        }
+    }
+
+    fun confirmCustom(
+        testParameters: TestParameters,
+        populateCustomLpmFields: () -> Unit = {},
+        verifyCustomLpmFields: () -> Unit = {}
+    ) {
+        setup(testParameters)
+        launchCustom()
+
+        selectors.paymentSelection.click()
+
+        val fieldPopulator = FieldPopulator(
+            selectors,
+            testParameters,
+            populateCustomLpmFields,
+            verifyCustomLpmFields
+        )
+        fieldPopulator.populateFields()
+
+        Espresso.onIdle()
+        composeTestRule.waitForIdle()
+
+        pressContinue()
+
+        pressMultiStepSelect()
+
+        Espresso.onIdle()
+        composeTestRule.waitForIdle()
+
+        fieldPopulator.verifyFields()
+
+        teardown()
+    }
+
+    private fun pressMultiStepSelect() {
+        selectors.multiStepSelect.click()
+        waitForNotPlaygroundActivity()
+    }
+
+    private fun pressContinue() {
+        selectors.continueButton.apply {
+            scrollTo()
+            click()
+        }
+
+        waitForPlaygroundActivity()
+    }
 
     /**
      * This will open the payment sheet complete flow from the playground with a new or
@@ -50,17 +112,13 @@ class PlaygroundTestDriver(
         testParameters: TestParameters,
         populateCustomLpmFields: () -> Unit = {}
     ) {
-        this.selectors = Selectors(device, composeTestRule, testParameters)
-        this.testParameters = testParameters
-
-        registerListeners()
-        launchComplete(selectors)
-
-        // PaymentSheetActivity is now on screen
-        callbackLock.acquire()
+        setup(testParameters)
+        launchComplete()
 
         selectors.paymentSelection.click()
 
+        // This takes a screenshot so that translation strings of placeholders
+        // and labels and design can all be verified
         takeScreenShot(
             fileName = "${selectors.baseScreenshotFilenamePrefix}-beforeText",
             testParameters.takeScreenshotOnLpmLoad
@@ -69,9 +127,11 @@ class PlaygroundTestDriver(
         FieldPopulator(
             selectors,
             testParameters,
-            populateCustomLpmFields
+            populateCustomLpmFields,
         ).populateFields()
 
+        // This takes a screenshot so that design and style can be verified after
+        // user input is entered.
         takeScreenShot(
             fileName = "${selectors.baseScreenshotFilenamePrefix}-afterText",
             testParameters.takeScreenshotOnLpmLoad
@@ -84,14 +144,38 @@ class PlaygroundTestDriver(
             testParameters.useBrowser
         )
 
+        pressBuy()
+
+        doAuthorization()
+
+        teardown()
+    }
+
+    private fun pressBuy() {
         selectors.buyButton.apply {
             scrollTo()
             click()
         }
+    }
 
-        doAuthorization()
+    /**
+     * Here we wait for an activity different from the playground to be in view.  We
+     * don't specifically look for PaymentSheetActivity or PaymentOptionsActivity because
+     * that would require exposing the activities publicly.
+     */
+    private fun waitForNotPlaygroundActivity() {
+        while (currentActivity[0] is PaymentSheetPlaygroundActivity) {
+            TimeUnit.MILLISECONDS.sleep(250)
+        }
+    }
 
-        callbackLock.release()
+    /**
+     * Here we wait for the Playground to come back into view.
+     */
+    private fun waitForPlaygroundActivity() {
+        while (currentActivity[0] !is PaymentSheetPlaygroundActivity) {
+            TimeUnit.MILLISECONDS.sleep(250)
+        }
     }
 
     private fun verifyDeviceSupportsTestAuthorization(
@@ -116,43 +200,12 @@ class PlaygroundTestDriver(
         } ?: installedBrowsers.first()
     }
 
-    internal fun registerListeners() {
-        val launchPlayground = Semaphore(1)
-        launchPlayground.acquire()
-        // Setup the playground for scenario, and launch it.  We use the playground
-        // so we don't have to implement another route to create a payment intent,
-        // the challenge is that we don't have access to the activity or it's viewmodels
-        val scenario = ActivityScenario.launch(PaymentSheetPlaygroundActivity::class.java)
-        scenario.onActivity { activity ->
-
-            IdlingPolicies.setIdlingResourceTimeout(45, TimeUnit.SECONDS)
-            IdlingPolicies.setMasterPolicyTimeout(45, TimeUnit.SECONDS)
-
-            IdlingRegistry.getInstance().register(
-                activity.getMultiStepIdlingResource(),
-                activity.getSingleStepIdlingResource(),
-                TransitionFragmentResource.getSingleStepIdlingResource()
-            )
-
-            // Observe the result of the action
-            activity.viewModel.status.observeForever {
-                resultValue = it
-                callbackLock.release()
-
-                IdlingRegistry.getInstance().unregister(
-                    activity.getMultiStepIdlingResource(),
-                    activity.getSingleStepIdlingResource(),
-                    TransitionFragmentResource.getSingleStepIdlingResource()
-                )
-            }
-            launchPlayground.release()
-        }
-
-        launchPlayground.acquire()
-        launchPlayground.release()
+    private fun monitorCurrentActivity(application: Application) {
+        this.application = application
+        application.registerActivityLifecycleCallbacks(activityLifecycleCallbacks)
     }
 
-    internal fun launchComplete(selectors: Selectors) {
+    private fun setConfiguration(selectors: Selectors) {
         // Could consider setting these preferences instead of clicking
         // if it is faster (possibly 1-2s)
         selectors.customer.click()
@@ -166,8 +219,25 @@ class PlaygroundTestDriver(
         // Can't guarantee that google pay will be on the phone
         selectors.googlePayState.click()
 
-        EspressoLabelIdButton(R.string.reload_paymentsheet).click()
-        EspressoLabelIdButton(R.string.checkout_complete).click()
+    }
+
+    internal fun launchComplete() {
+        selectors.reload.click()
+        selectors.complete.click()
+
+        // PaymentSheetActivity is now on screen
+        waitForNotPlaygroundActivity()
+    }
+
+    private fun launchCustom() {
+        selectors.reload.click()
+        Espresso.onIdle()
+        selectors.composeTestRule.waitForIdle()
+
+        selectors.multiStepSelect.click()
+
+        // PaymentOptionsActivity is now on screen
+        waitForNotPlaygroundActivity()
     }
 
     private fun doAuthorization() {
@@ -229,7 +299,7 @@ class PlaygroundTestDriver(
         if (testParameters.authorizationAction == AuthorizeAction.Authorize
             || testParameters.authorizationAction == null
         ) {
-            callbackLock.acquire()
+            waitForPlaygroundActivity()
             assertThat(resultValue).isEqualTo(
                 PaymentSheetResult.Completed.toString()
             )
@@ -247,5 +317,50 @@ class PlaygroundTestDriver(
 
             capture.process(setOf(basicScreenCaptureProcessor))
         }
+    }
+
+    internal fun setup(testParameters: TestParameters) {
+        this.testParameters = testParameters
+        this.selectors = Selectors(device, composeTestRule, testParameters)
+
+        val launchPlayground = Semaphore(1)
+        launchPlayground.acquire()
+
+        // Setup the playground for scenario, and launch it.  We use the playground
+        // so we don't have to implement another route to create a payment intent,
+        // the challenge is that we don't have access to the activity or it's viewmodels
+        val scenario = ActivityScenario.launch(PaymentSheetPlaygroundActivity::class.java)
+        scenario.onActivity { activity ->
+
+            monitorCurrentActivity(activity.application)
+
+            IdlingPolicies.setIdlingResourceTimeout(45, TimeUnit.SECONDS)
+            IdlingPolicies.setMasterPolicyTimeout(45, TimeUnit.SECONDS)
+
+            IdlingRegistry.getInstance().register(
+                activity.getMultiStepReadyIdlingResource(),
+                activity.getSingleStepReadyIdlingResource(),
+            )
+
+            // Observe the result of the PaymentSheet completion
+            activity.viewModel.status.observeForever {
+                resultValue = it
+
+                IdlingRegistry.getInstance().unregister(
+                    activity.getMultiStepReadyIdlingResource(),
+                    activity.getSingleStepReadyIdlingResource(),
+                )
+            }
+            launchPlayground.release()
+        }
+
+        launchPlayground.acquire()
+        launchPlayground.release()
+
+        setConfiguration(selectors)
+    }
+
+    internal fun teardown() {
+        application?.unregisterActivityLifecycleCallbacks(activityLifecycleCallbacks)
     }
 }