+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/test/core/ui/PaymentSelection.kt
@@ -8,6 +8,7 @@ import androidx.compose.ui.test.junit4.ComposeTestRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
+import androidx.compose.ui.test.performScrollTo
 import androidx.compose.ui.test.performScrollToNode
 import androidx.test.platform.app.InstrumentationRegistry
 import com.stripe.android.paymentsheet.TEST_TAG_LIST
@@ -17,6 +18,7 @@ class PaymentSelection(val composeTestRule: ComposeTestRule, @StringRes val labe
         val resource = InstrumentationRegistry.getInstrumentation().targetContext.resources
         composeTestRule.onNodeWithTag(TEST_TAG_LIST, true)
             .performScrollToNode(hasText(resource.getString(label)))
+        composeTestRule.waitForIdle()
         composeTestRule
             .onNodeWithText(resource.getString(label))
             .assertIsDisplayed()