+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/activity/PaymentSheetPlaygroundActivity.kt
@@ -89,10 +89,10 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
     private lateinit var flowController: PaymentSheet.FlowController
 
     @Nullable
-    private var multiStepUIIdlingResource: CountingIdlingResource? = null
+    private var multiStepUIReadyIdlingResource: CountingIdlingResource? = null
 
     @Nullable
-    private var singleStepUIIdlingResource: CountingIdlingResource? = null
+    private var singleStepUIReadyIdlingResource: CountingIdlingResource? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -158,11 +158,10 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
         viewModel.inProgress.observe(this) {
             viewBinding.progressBar.isInvisible = !it
             if (it) {
-                singleStepUIIdlingResource?.increment()
-                multiStepUIIdlingResource?.increment()
+                singleStepUIReadyIdlingResource?.increment()
+                multiStepUIReadyIdlingResource?.increment()
             } else {
-                singleStepUIIdlingResource?.decrement()
-                multiStepUIIdlingResource?.decrement()
+                singleStepUIReadyIdlingResource?.decrement()
             }
         }
 
@@ -176,22 +175,6 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
         }
 
         disableViews()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        val (customer, link, googlePay, currency, mode, setShippingAddress, setDefaultBillingAddress, setAutomaticPaymentMethods, setDelayedPaymentMethods) = viewModel.getSavedToggleState()
-        setToggles(
-            customer,
-            link,
-            googlePay,
-            currency,
-            mode,
-            setShippingAddress,
-            setDefaultBillingAddress,
-            setAutomaticPaymentMethods,
-            setDelayedPaymentMethods
-        )
     }
 
     override fun onPause() {
@@ -337,6 +320,7 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
         if (success) {
             viewBinding.paymentMethod.isClickable = true
             onPaymentOption(flowController.getPaymentOption())
+            multiStepUIReadyIdlingResource?.decrement()
         } else {
             viewModel.status.value =
                 "Failed to configure PaymentSheetFlowController: ${error?.message}"
@@ -369,26 +353,28 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
     }
 
     /**
-     * Only called from test, creates and returns a new [SimpleIdlingResource].
+     * Only called from test, creates and returns a [IdlingResource].
      */
-
     @VisibleForTesting
     @NonNull
-    fun getMultiStepIdlingResource(): IdlingResource? {
-        if (multiStepUIIdlingResource == null) {
-            multiStepUIIdlingResource = CountingIdlingResource("multiStepUIIdlingResource")
+    fun getMultiStepReadyIdlingResource(): IdlingResource? {
+        if (multiStepUIReadyIdlingResource == null) {
+            multiStepUIReadyIdlingResource =
+                CountingIdlingResource("multiStepUIReadyIdlingResource")
         }
-        return multiStepUIIdlingResource
+        return multiStepUIReadyIdlingResource
     }
 
     @VisibleForTesting
     @NonNull
-    fun getSingleStepIdlingResource(): IdlingResource? {
-        if (singleStepUIIdlingResource == null) {
-            singleStepUIIdlingResource = CountingIdlingResource("singleStepUIIdlingResource")
+    fun getSingleStepReadyIdlingResource(): IdlingResource? {
+        if (singleStepUIReadyIdlingResource == null) {
+            singleStepUIReadyIdlingResource =
+                CountingIdlingResource("singleStepUIReadyIdlingResource")
         }
-        return singleStepUIIdlingResource
+        return singleStepUIReadyIdlingResource
     }
+
 
     companion object {
         private const val merchantName = "Example, Inc."