+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/model/SupportedPaymentMethod.kt
@@ -16,8 +16,6 @@ import com.stripe.android.paymentsheet.forms.AfterpayClearpayRequirement
 import com.stripe.android.paymentsheet.forms.AuBecsDebitRequirement
 import com.stripe.android.paymentsheet.forms.BancontactRequirement
 import com.stripe.android.paymentsheet.forms.CardRequirement
-import com.stripe.android.ui.core.forms.CardForm
-import com.stripe.android.ui.core.forms.CardParamKey
 import com.stripe.android.paymentsheet.forms.Delayed
 import com.stripe.android.paymentsheet.forms.EpsRequirement
 import com.stripe.android.paymentsheet.forms.GiropayRequirement
@@ -35,31 +33,19 @@ import com.stripe.android.paymentsheet.forms.USBankAccountRequirement
 import com.stripe.android.ui.core.elements.LayoutFormDescriptor
 import com.stripe.android.ui.core.elements.LayoutSpec
 import com.stripe.android.ui.core.forms.AffirmForm
-import com.stripe.android.ui.core.forms.AffirmParamKey
 import com.stripe.android.ui.core.forms.AfterpayClearpayForm
-import com.stripe.android.ui.core.forms.AfterpayClearpayParamKey
-import com.stripe.android.ui.core.forms.BancontactForm
-import com.stripe.android.ui.core.forms.BancontactParamKey
-import com.stripe.android.ui.core.forms.EpsForm
-import com.stripe.android.ui.core.forms.EpsParamKey
-import com.stripe.android.ui.core.forms.GiropayForm
-import com.stripe.android.ui.core.forms.GiropayParamKey
-import com.stripe.android.ui.core.forms.IdealForm
-import com.stripe.android.ui.core.forms.IdealParamKey
-import com.stripe.android.ui.core.forms.KlarnaForm
-import com.stripe.android.ui.core.forms.KlarnaParamKey
-import com.stripe.android.ui.core.forms.P24Form
-import com.stripe.android.ui.core.forms.P24ParamKey
-import com.stripe.android.ui.core.forms.PaypalForm
-import com.stripe.android.ui.core.forms.PaypalParamKey
-import com.stripe.android.ui.core.forms.SepaDebitForm
-import com.stripe.android.ui.core.forms.SepaDebitParamKey
-import com.stripe.android.ui.core.forms.SofortForm
-import com.stripe.android.ui.core.forms.SofortParamKey
 import com.stripe.android.ui.core.forms.AuBecsDebitForm
-import com.stripe.android.ui.core.forms.AuBecsDebitParamKey
+import com.stripe.android.ui.core.forms.BancontactForm
+import com.stripe.android.ui.core.forms.CardForm
+import com.stripe.android.ui.core.forms.EpsForm
+import com.stripe.android.ui.core.forms.GiropayForm
+import com.stripe.android.ui.core.forms.IdealForm
+import com.stripe.android.ui.core.forms.KlarnaForm
+import com.stripe.android.ui.core.forms.P24Form
+import com.stripe.android.ui.core.forms.PaypalForm
+import com.stripe.android.ui.core.forms.SepaDebitForm
+import com.stripe.android.ui.core.forms.SofortForm
 import com.stripe.android.ui.core.forms.USBankAccountForm
-import com.stripe.android.ui.core.forms.USBankAccountParamKey
 import kotlinx.parcelize.Parcelize
 
 /**
@@ -91,11 +77,6 @@ sealed class SupportedPaymentMethod(
     private val requirement: PaymentMethodRequirements,
 
     /**
-     * This is a map of the fields in payment_method_options.  See [this](https://stripe.com/docs/api/payment_intents/create#create_payment_intent-payment_method_options) for details.
-     */
-    val paramKey: MutableMap<String, Any?>,
-
-    /**
      * This describes how the UI should look.
      */
     val formSpec: LayoutSpec,
@@ -107,7 +88,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_card,
         R.drawable.stripe_ic_paymentsheet_pm_card,
         CardRequirement,
-        CardParamKey,
         CardForm
     )
 
@@ -118,7 +98,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_bancontact,
         R.drawable.stripe_ic_paymentsheet_pm_bancontact,
         BancontactRequirement,
-        BancontactParamKey,
         BancontactForm
     )
 
@@ -129,7 +108,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_sofort,
         R.drawable.stripe_ic_paymentsheet_pm_klarna,
         SofortRequirement,
-        SofortParamKey,
         SofortForm
     )
 
@@ -140,7 +118,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_ideal,
         R.drawable.stripe_ic_paymentsheet_pm_ideal,
         IdealRequirement,
-        IdealParamKey,
         IdealForm
     )
 
@@ -151,7 +128,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_sepa_debit,
         R.drawable.stripe_ic_paymentsheet_pm_sepa_debit,
         SepaDebitRequirement,
-        SepaDebitParamKey,
         SepaDebitForm
     )
 
@@ -162,7 +138,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_eps,
         R.drawable.stripe_ic_paymentsheet_pm_eps,
         EpsRequirement,
-        EpsParamKey,
         EpsForm
     )
 
@@ -173,7 +148,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_p24,
         R.drawable.stripe_ic_paymentsheet_pm_p24,
         P24Requirement,
-        P24ParamKey,
         P24Form
     )
 
@@ -184,7 +158,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_giropay,
         R.drawable.stripe_ic_paymentsheet_pm_giropay,
         GiropayRequirement,
-        GiropayParamKey,
         GiropayForm
     )
 
@@ -195,7 +168,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_afterpay_clearpay,
         R.drawable.stripe_ic_paymentsheet_pm_afterpay_clearpay,
         AfterpayClearpayRequirement,
-        AfterpayClearpayParamKey,
         AfterpayClearpayForm
     )
 
@@ -206,7 +178,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_klarna,
         R.drawable.stripe_ic_paymentsheet_pm_klarna,
         KlarnaRequirement,
-        KlarnaParamKey,
         KlarnaForm
     )
 
@@ -217,7 +188,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_paypal,
         R.drawable.stripe_ic_paymentsheet_pm_paypal,
         PaypalRequirement,
-        PaypalParamKey,
         PaypalForm
     )
 
@@ -228,7 +198,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_affirm,
         R.drawable.stripe_ic_paymentsheet_pm_affirm,
         AffirmRequirement,
-        AffirmParamKey,
         AffirmForm
     )
 
@@ -239,7 +208,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_au_becs_debit,
         R.drawable.stripe_ic_paymentsheet_pm_bank,
         AuBecsDebitRequirement,
-        AuBecsDebitParamKey,
         AuBecsDebitForm
     )
 
@@ -250,7 +218,6 @@ sealed class SupportedPaymentMethod(
         R.string.stripe_paymentsheet_payment_method_us_bank_account,
         R.drawable.stripe_ic_paymentsheet_pm_bank,
         USBankAccountRequirement,
-        USBankAccountParamKey,
         USBankAccountForm
     )